Add to readme, additional yaml checkin
