+++ src/examples/learn_quant/conf/grammar/typed_rules/set_primitives.yaml
@@ -1,0 +1,1 @@
+module_path: learn_quant.set_primitives
