+++ src/examples/learn_quant/conf/recipe/4_4_5_xi.yaml
@@ -1,0 +1,29 @@
+# @package _global_
+defaults:
+  - //Users/radmajik/Documents/UWLing/altk/src/examples/learn_quant/conf/grammar/base@grammar
+  - //Users/radmajik/Documents/UWLing/altk/src/examples/learn_quant/conf/universe/base@universe
+  - _self_
+
+name: base
+
+grammar: 
+  depth: 5
+  indices: false
+  weight: 2.0
+
+universe:
+  m_size: 4
+  x_size: 4
+  inclusive_universes: ${universe.inclusive_universes}
+
+measures:
+  expressions:
+    - subset_eq(A, B)
+  monotonicity:
+    debug: false
+    universe_filter:
+      - 3
+      - 4
+    direction:
+      - all
+
