+++ src/examples/learn_quant/monotonicity.py
@@ -3,7 +3,6 @@ from ultk.language.grammar import GrammaticalExpression
 from ultk.language.language import Language, aggregate_expression_complexity
 from ultk.language.semantics import Meaning
 
-from learn_quant.meaning import create_universe
 from learn_quant.util import read_expressions
 from learn_quant.quantifier import QuantifierUniverse, QuantifierModel
 from learn_quant.grammar import add_indices, get_indices_tag, QuantifierGrammar
@@ -26,6 +25,36 @@ from omegaconf import DictConfig, OmegaConf
 # e.g.:
 # python -m learn_quant.monotonicity recipe=test_monotonicity
 # python -m learn_quant.monotonicity recipe=4_4_5_xi grammar.indices=false
+
+def create_universe(m_size: int, x_size: int) -> QuantifierUniverse:
+    """
+    Create a quantifier universe based on the given parameters.
+    All references are quantifier models, which are data classes that represent a relation between sets A, B, and M.
+
+    Args:
+        m_size (int): The size of the m set.
+        x_size (int): The size of the x set.
+
+    Returns:
+        QuantifierUniverse: The created quantifier universe.
+    """
+
+    possible_quantifiers = []
+
+    for combination in combinations_with_replacement([0, 1, 2, 3], r=m_size):
+        combo = list(combination) + [4] * (x_size - m_size)
+        permutations_for_combo = set(permutations(combo, r=len(combo)))
+        possible_quantifiers.extend(permutations_for_combo)
+    possible_quantifiers_name = set(
+        ["".join([str(j) for j in i]) for i in possible_quantifiers]
+    )
+
+    quantifier_models = set()
+    for name in possible_quantifiers_name:
+        quantifier_models.add(QuantifierModel(name=name))
+
+    return QuantifierUniverse(referents=tuple(quantifier_models), m_size=m_size, x_size=x_size)
+
 
 def binary_to_int(arr):
     """ Converts a 2-D numpy array of 1s and 0s into integers, assuming each
